refactor: apply code review comments
