+++ Source/Kysect.Shreks.WebApi/TestEnvConfiguration.cs
@@ -1,0 +1,5 @@
+public class TestEnvConfiguration
+{
+    public string Organization { get; init; }
+    public List<string> Users { get; init; }
+}
